+++ source/common/modules/preload/index.ts
@@ -14,7 +14,7 @@
  * END HEADER
  */
 
-import { contextBridge, ipcRenderer, clipboard } from 'electron'
+import { contextBridge, ipcRenderer } from 'electron'
 
 // PREPARATION: Since we have multiple editor panes and all of them need to
 // listen to a few events, we need to ramp up some of the channels' max
@@ -71,84 +71,6 @@ contextBridge.exposeInMainWorld(
     }
   }
 )
-
-// Expose the subset of clipboard functions which we use
-contextBridge.exposeInMainWorld('clipboard', {
-  readText: function () {
-    return clipboard.readText()
-  },
-  readHTML: function () {
-    return clipboard.readHTML()
-  },
-  readRTF: function () {
-    return clipboard.readRTF()
-  },
-  hasImage: function () {
-    // NOTE: We cannot send NativeImages across the context bridge,
-    // so we have to do it the hard way here.
-    return !clipboard.readImage().isEmpty()
-  },
-  getImageData: function () {
-    // NOTE: This function is used only in the Paste-Image dialog in order to
-    // show a preview of the image and populate the dialog with the necessary
-    // information.
-    const image = clipboard.readImage()
-    const size = image.getSize() // First get the original size
-    const aspect = image.getAspectRatio() // Then the aspect
-    const dataUrl = image.resize({ height: 200 }).toDataURL()
-
-    return {
-      size,
-      aspect,
-      dataUrl
-    }
-  },
-  write: function (data: Electron.Data) {
-    // We cannot transfer images, so make sure the function is safe to call
-    return clipboard.write({
-      text: data.text,
-      html: data.html,
-      rtf: data.rtf
-    })
-  },
-  writeText: function (text: string) {
-    return clipboard.writeText(text)
-  },
-  hasSelectionClipboard: function () {
-    if (process.platform !== 'linux') {
-      return false
-    }
-
-    if (clipboard.readText('selection') !== '') {
-      return true
-    }
-
-    if (clipboard.readHTML('selection') !== '') {
-      return true
-    }
-
-    return false
-  },
-  /**
-   * Returns the plain text and HTML contents of the selection clipboard on
-   * linux.
-   *
-   * @return  {{text: string, html: string}}}  Returns an object containing HTML and text contents
-   */
-  getSelectionClipboard: function () {
-    if (process.platform !== 'linux') {
-      return {
-        text: '',
-        html: ''
-      }
-    } else {
-      return {
-        text: clipboard.readText('selection'),
-        html: clipboard.readHTML('selection')
-      }
-    }
-  }
-})
 
 // Expose the subset of process properties we need
 contextBridge.exposeInMainWorld('process', {